% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.decomp.R
\docType{package}
\name{causal.decomp}
\alias{causal.decomp}
\title{causal.decomp: Causal Decomposition Analysis.}
\description{
The causal.decomp package provides four important functions:
mmi, smi, pocr, and sensitivity.
}
