% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.sensitivity}
\title{Visualize sensitivity Objects}
\usage{
\method{plot}{sensitivity}(x, xlim = c(0, 0.3), ylim = c(0, 0.3), ...)
}
\arguments{
\item{x}{an object of class \code{sensitivity}.}

\item{xlim}{limits of the x-axis of contour plots.}

\item{ylim}{limits of the y-axis of contour plots.}

\item{...}{Other arguments for future usage.}
}
\description{
S3 methods visualizing results for some objects generated by \code{sensitivity}.
}
